requirements.txt
streamlit
google-generativeai
streamlit-autorefresh